requirements.txt
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.10
bcrypt==4.0.1
black==25.1.0
cfgv==3.4.0
click==8.2.1
dill==0.4.0
distlib==0.3.9
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.12
filelock==3.18.0
flake8==7.2.0
greenlet==3.2.3
h11==0.16.0
identify==2.6.12
idna==3.10
iniconfig==2.1.0
isort==6.0.1
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
psycopg2==2.9.10
pyasn1==0.6.1
pycodestyle==2.13.0
pydantic==2.11.5
pydantic_core==2.33.2
pyflakes==3.3.2
Pygments==2.19.1
pylint==3.3.7
pytest==8.4.0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tomli==2.2.1
tomlkit==0.13.3
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
virtualenv==20.31.2
